+++ NEWS.md
@@ -1,9 +1,10 @@
 # timevis 0.5.*
 
-TODO
+TODO date
 
 - Add `timezone` parameter to `timevis()` - supports showing the timeline in a different timezone than your local machine (#67)
 - Add `setCustomTime()` and `setCurrentTime()` functions (#20)
+- Fixed bug: using `setSelection()` did not trigger the Shiny selected input (#82)
 - Add documentation about how to add custom style to timevis (#45)
 - Add documentation tab to the demo app